move calendar cell to different component
