+++ client/src/components/Calendar/CalendarSetup.tsx
@@ -2,6 +2,13 @@ import { dateFnsLocalizer } from 'react-big-calendar'
 import { format, parse, startOfWeek, getDay } from 'date-fns'
 import enUS from 'date-fns/locale/en-US'
 
+export type TaskSummary = {
+  [date: string]: {
+    assigned: number
+    total: number
+  }
+}
+
 const locales = {
   'en-US': enUS,
 }
